+++ CHANGES.md
@@ -5,6 +5,7 @@
 ## unreleased
 
 * Fixed: replace `msl:tokenizedString` with `msl:variables` in the `variable` attribute of `range`
+* Added `recalibrateReference` attribute to `measurementStartDate` and `measurementStopDate` elements.
 * Added the following to `quantityEnumerationString`: Spectral Power Responsivity
 * Added the following to `staffEnumerationString`: Claude Meffan, Charlotte Hayton, Johannes Kaiser
 * Added the following to `currencyEnumerationString`: CHF